feat(ci): configured only one release CI file
The release workflow will run from a dispatch event that is manually activated.
Depending on which branch the workflow is triggered, the semantic release will generate a release, if it occurs in the
main branch (default) or a prerelease if it occurs in the alpha branch.
The despatch event is automatically triggered when a push to the alpha branch occurs and the test workflow returns success.